Checking the FULL name (check_fio): After entering the full name, the bot checks its presence in the users_info table. If the full name is already used, the application is rejected. If not, the user_id is associated with this full name, blocking it for other users.
Adding a description of the problem (ask_for_message and save_request): Now the user enters a description of the problem, which is saved in a new message column and sent to the operators.

Division of applications by tables: Depending on the type of application (by documents, by number, by application), data is saved in the corresponding tables (requests_documents, requests_number, requests_request).

Saving the time of sending and responding: Columns time_submitted (automatically filled in when inserting a record) and time_answered (filled in when the operator responds to the request) have been added to the tables.

Processing operator responses: When the operator responds, the bot determines which table the request is in by operator_message_id, forwards the response to the user and saves the response time.